MPL CNN for IFT_6390_Hw3
In this part of the homework, you are going to test your implementation on the SVHN dataset.
You should download the dataset from here. The SVHN dataset that we are using consists of 90,000 32 × 32 coloured images in 10 classes. We are going to use 67,000 images for training, 3000 images for validation, and 20,000 images for testing. Each image, originally represented with a 32 × 32 × 3 matrix, has been flattened, meaning that each example is represented with an array of size 3072. The file you downloaded is a pickle file, that can be processed with the code provided to you in NN.init. If you initialize your NN class by providing the path to the pickle file you downloaded in the parameter datapath, you will end up with the variables self.train, self.valid, self.test containing the dataset. Each of these three variables is a tuple. For example, self.valid[0] is a matrix of size 3000 × 3072, and self.valid[1] is an array of size 3000 containing the labels of the 3000 validation set images, from 0 to 9. Please double check that the six arrays have the right size!
The answers to the questions of this section, along with the required figures, should be written in a report you will submit to Gradescope as the practical part of the homework (seperatly from the Theoretic part).